refactor: move common methods to new util
